Trim fclShopMisc 003f99d0 stack and loop widths

+++ src/Event/Fcl/fclShopMisc.c
@@ -7738,7 +7738,7 @@ u64 FUN_003f99d0(u32 param_1,u32 param_2)
 
   int lVar6;
 
-  u32 uVar7;
+  int uVar7;
 
   u8 auStack_20 [0x20];
 
@@ -7760,7 +7760,7 @@ u64 FUN_003f99d0(u32 param_1,u32 param_2)
 
       if (((*(u32 *)(auStack_20 + 0x1c) & 0x1000) != 0) &&
          (sVar3 = datGetEquipmentIdx(1,(u16)(*(u32 *)(auStack_20 + 0x1c) >> 8) & 0xf),
-          uVar7 == (long)sVar3)) {
+          uVar7 == sVar3)) {
 
         uVar2 = *puVar1;
